type fix

+++ src/main/java/org/geppetto/model/swc/format/SWCReader.java
@@ -104,7 +104,7 @@ public class SWCReader
 					x = lineValues[2];
 					y = lineValues[3];
 					z = lineValues[4];
-					if ((int) lineValues[5] > 0){ // enforsing a minimum radius of 1
+					if (lineValues[5] > 0.0){ // enforsing a minimum radius of 1
 						radius = lineValues[5];
 					}else{
 						radius = 1.0;
